+++ Frameworks/AVFoundation.framework/AVMetadataEnumerator.h
@@ -6,8 +6,8 @@
 
 @interface AVMetadataEnumerator : NSEnumerator {
     NSEnumerator *_arrayEnumerator;
-    long long _count;
-    long long _index;
+    long _count;
+    long _index;
     id _key;
     NSString *_keySpace;
     NSLocale *_locale;
