+++ Frameworks/AVKit.framework/AVPlayerView.h
@@ -3,15 +3,15 @@
  */
 
 @interface AVPlayerView : UIView {
-    long long _viewWindowState;
+    int _viewWindowState;
 }
 
-@property(readonly) long long viewWindowState;
+@property(readonly) int viewWindowState;
 
 - (void)_updateViewWindowState;
 - (void)didMoveToWindow;
-- (void)setFrame:(struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })arg1;
-- (long long)viewWindowState;
+- (void)setFrame:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1;
+- (int)viewWindowState;
 - (void)willMoveToWindow:(id)arg1;
 
 @end
