+++ Frameworks/AddressBookUI.framework/ABPersonTableTinyActionCell.h
@@ -12,24 +12,24 @@
 }
 
 @property(retain) ABPersonTableActionDataSource * actionDataSource;
-@property(readonly) unsigned long long actionsCount;
+@property(readonly) unsigned int actionsCount;
 @property <ABPersonTableTinyActionDelegate> * delegate;
 @property(retain) <ABStyleProvider> * styleProvider;
 
 - (void)_resizeInnerContentView;
 - (id)_tinyActionContentView;
 - (id)actionDataSource;
-- (unsigned long long)actionsCount;
+- (unsigned int)actionsCount;
 - (void)dealloc;
 - (id)delegate;
-- (id)initWithStyle:(long long)arg1 reuseIdentifier:(id)arg2;
+- (id)initWithStyle:(int)arg1 reuseIdentifier:(id)arg2;
 - (void)layoutSubviews;
 - (void)refreshActions;
 - (void)setActionDataSource:(id)arg1;
 - (void)setBackgroundView:(id)arg1;
 - (void)setDelegate:(id)arg1;
 - (void)setStyleProvider:(id)arg1;
-- (struct CGSize { double x1; double x2; })sizeThatFits:(struct CGSize { double x1; double x2; })arg1;
+- (struct CGSize { float x1; float x2; })sizeThatFits:(struct CGSize { float x1; float x2; })arg1;
 - (id)styleProvider;
 
 @end
