+++ Frameworks/AddressBookUI.framework/ABRepeatingGradientSeparatorView.h
@@ -6,25 +6,25 @@
 
 @interface ABRepeatingGradientSeparatorView : UIView {
     UIColor *_endColor;
-    long long _lineCount;
+    int _lineCount;
     UIColor *_startColor;
 }
 
 @property(retain) UIColor * endColor;
 @property(readonly) NSArray * gradientColors;
-@property long long lineCount;
+@property int lineCount;
 @property(retain) UIColor * startColor;
 
 - (void)_updateGradients;
 - (void)dealloc;
 - (id)endColor;
 - (id)gradientColors;
-- (id)initWithFrame:(struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })arg1;
-- (struct CGSize { double x1; double x2; })intrinsicContentSize;
+- (id)initWithFrame:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1;
+- (struct CGSize { float x1; float x2; })intrinsicContentSize;
 - (void)layoutSubviews;
-- (long long)lineCount;
+- (int)lineCount;
 - (void)setEndColor:(id)arg1;
-- (void)setLineCount:(long long)arg1;
+- (void)setLineCount:(int)arg1;
 - (void)setStartColor:(id)arg1;
 - (id)startColor;
