+++ Frameworks/CoreAudioKit.framework/CAIAANodeContainer.h
@@ -3,18 +3,18 @@
  */
 
 @interface CAIAANodeContainer : UIScrollView {
-    long long _numPages;
+    int _numPages;
 }
 
-@property long long numPages;
+@property int numPages;
 
 + (void)initialize;
 
-- (double)contentWidth;
-- (id)initWithFrame:(struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })arg1;
+- (float)contentWidth;
+- (id)initWithFrame:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1;
 - (void)layoutSubviews;
-- (double)minWidth;
-- (long long)numPages;
-- (void)setNumPages:(long long)arg1;
+- (float)minWidth;
+- (int)numPages;
+- (void)setNumPages:(int)arg1;
 
 @end
