+++ Frameworks/CoreData.framework/NSKnownKeysMappingStrategy.h
@@ -6,19 +6,19 @@
 }
 
 @property(readonly) id* keys;
-@property(readonly) unsigned long long length;
+@property(readonly) unsigned int length;
 
 + (id)alloc;
 + (id)allocWithZone:(struct _NSZone { }*)arg1;
 
-- (unsigned long long)countByEnumeratingWithState:(struct { unsigned long long x1; id *x2; unsigned long long *x3; unsigned long long x4[5]; }*)arg1 objects:(id*)arg2 count:(unsigned long long)arg3;
+- (unsigned int)countByEnumeratingWithState:(struct { unsigned long x1; id *x2; unsigned long *x3; unsigned long x4[5]; }*)arg1 objects:(id*)arg2 count:(unsigned int)arg3;
 - (void)encodeWithCoder:(id)arg1;
-- (unsigned long long)fastIndexForKnownKey:(id)arg1;
-- (unsigned long long)indexForKey:(id)arg1;
+- (unsigned int)fastIndexForKnownKey:(id)arg1;
+- (unsigned int)indexForKey:(id)arg1;
 - (id)initForKeys:(id)arg1;
-- (id)initForKeys:(id*)arg1 count:(unsigned long long)arg2;
+- (id)initForKeys:(id*)arg1 count:(unsigned int)arg2;
 - (id)initWithCoder:(id)arg1;
 - (id*)keys;
-- (unsigned long long)length;
+- (unsigned int)length;
 
 @end
